Clean up YieldProcessorNormalized (#14739)
Move YieldProcessorNormalized into separate files

Clean up YieldProcessorNormalized
